+++ src/Handler.php
@@ -29,6 +29,25 @@ class Handler
     protected $previousExceptionHandler;
 
     /**
+     * A bit of reserved memory to ensure we are able to increase the memory
+     * limit on an OOM.
+     *
+     * We can't reserve all of the memory that we need to send OOM reports
+     * because this would have a big overhead on every request, instead of just
+     * on shutdown in requests with errors.
+     *
+     * @var string|null
+     */
+    private $reservedMemory;
+
+    /**
+     * A regex that matches PHP OOM errors.
+     *
+     * @var string
+     */
+    private $oomRegex = '/^Allowed memory size of (\d+) bytes exhausted \(tried to allocate \d+ bytes\)/';
+
+    /**
      * Whether the shutdown handler will run.
      *
      * This is used to disable the shutdown handler in order to avoid double
@@ -136,6 +155,9 @@ class Handler
      */
     public function registerShutdownHandler()
     {
+        // Reserve some memory that we can free in the shutdown handler
+        $this->reservedMemory = str_repeat(' ', 1024 * 32);
+
         register_shutdown_function([$this, 'shutdownHandler']);
     }
 
@@ -267,6 +289,9 @@ class Handler
      */
     public function shutdownHandler()
     {
+        // Free the reserved memory to give ourselves some room to work
+        $this->reservedMemory = null;
+
         // If we're disabled, do nothing. This avoids reporting twice if the
         // exception handler is forcing the native PHP handler to run
         if (!self::$enableShutdownHandler) {
@@ -274,6 +299,17 @@ class Handler
         }
 
         $lastError = error_get_last();
+
+        // If this is an OOM and memory increase is enabled, bump the memory
+        // limit so we can report it
+        if ($lastError !== null
+            && $this->client->getMemoryLimitIncrease() !== null
+            && preg_match($this->oomRegex, $lastError['message'], $matches) === 1
+        ) {
+            $currentMemoryLimit = (int) $matches[1];
+
+            ini_set('memory_limit', $currentMemoryLimit + $this->client->getMemoryLimitIncrease());
+        }
 
         // Check if a fatal error caused this shutdown
         if (!is_null($lastError) && ErrorTypes::isFatal($lastError['type']) && !$this->client->getConfig()->shouldIgnoreErrorCode($lastError['type'])) {