+++ tests/Fakes/SomeException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Bugsnag\Tests\Fakes;
+
+use Exception;
+
+final class SomeException extends Exception
+{
+}