+++ tests/phpt/_prelude.php
@@ -6,14 +6,20 @@ use Bugsnag\Client;
 use Bugsnag\Configuration;
 use Bugsnag\Tests\phpt\Utilities\FakeGuzzle;
 
+date_default_timezone_set('UTC');
+
 $config = new Configuration('11111111111111111111111111111111');
 $config->setNotifyEndpoint('http://localhost/notify');
 $config->setSessionEndpoint('http://localhost/sessions');
 
 $guzzle = new FakeGuzzle();
 
-return new Client(
+$client = new Client(
     $config,
     null,
     $guzzle
 );
+
+$client->registerDefaultCallbacks();
+
+return $client;